New method of handling switching between coord systems.  Now the following rules are imposed:
User enters a target which is not a named one (eg 3C84) and selects the named target reference system:
A popup appears asking if they want to do this.  If they select "yes" then the name is wiped and the
named reference system tab is enabled.  If they select "no" then the name remains and the orbital elements
tab is enabled.

Users enters a named object (eg Venus) and selects RA/DEC coord sys.  User is presented with a popup warning
asking if they really want to do this.  If they select "yes" then the name is NOT cleared, but the RA/DEC tab
is enabled.  If they select "no" then the Named System tab is enabled.