Added the schedulable headers as requested. This is so that the data reduction can alarm when it detects that data does not conform to the requested bounds. Hopefully this will be a rare occasion, as the query tool selects only MSBs that are appropriate to the conditions, but it is possible that the situation will drift.
If someone is looking at this later: I apologise for the
brain-deadness of this code. We don't have a Java programmer right now
so I could only reproduce what was done for similar headers
previously, but there must be a more concise way of doing it. Will be
fixed by the new hire.