Change all Scan/Rasters to be called Scans.
Rasters scan a line and then return to the start of the row before
scanning the next, as this is inefficiant we commonly use what is referred
to as Boustrophedon, which means a line is scanned in one direction and
then the next line is scanned in the opposite direction. Boustrophedon is
not the only scanning strategy being used at JCMT, but is used in place
of Rastering.