Sub java.util.logging.Logger for Log4J alternative.
Makes for a cleaner build, the funtionality is close enough to use. This
helps greatly when doing webstart work, but also means I can build against
a vanilla JDK, so no frigging the $PATH at the summit.

There is a JACLogger class in the QT at the moment which should be moved
to the OT ASAP, however due to some crazy branching I'm doing it this way
until I've merged my branches.