Compress the Science Program template when replicating from catalogues.
Spotted by Erik Rosolowsky at University of British Columbia and reported
to the jcmtot mailing list. Requires work to the SOAP server to make it
work.

Essentially the problem is that the Science Program contains XML namespace
URLs and these cause the Perl SOAP::Lite to b0rk when using strict due to
dynamic exection of code that handles URLs. This is similar to the problem
we had with URLs in SpNotes.