Methanol lines requested by Jan ( 8 Jun 2009 )
This entry replaces a number of lines already in the catalogue with
versions with corrected transitions. The fact that the transitions were
incorrect is fairly worrying, hopefully I can knock together a utility
utilising SLAP to fix and update the catalogue. As I have a backlog of
lines to add I'm pushing on until someone complains.