Remove old Java libraries updated by JSky and update Web Start.
PAL from JSky3 has been removed in favour of the one already in
the repository, while they differ, we do not know which is which.

Web Start should now hopefully pull in the Starlink libraries,
this could potentially lead to errors and inconsistencies
between us and users, but we can correct the error should it
occur. It still may be better than shipping it ourselves.