Corrected a bug in the getTsys method
getTsys has a loop where it iterates through the tau values
supplied in the tau*.dat filenames and determines which files
should be used for the interpolation. getTsys had a bug such
that, if tau was supplied as a value between the first and second
tau values, the loop wouldn't recognize that fact and would end
up using the maximum tau values. Corrected the loop so now all
tau values are properly handled.