Begun adding reference arc plotting for FTS-2.
Added a class TpeReferenceArcFeature to plot 'sausages' showing
the area to be covered by a reference beam, as the source
changes in elevation.  Added this class to the JCMT configuration
file so that it is displayed for the FTS-2 observation.

In this version: the plotting routines work, the calculation
function gathers the data but does not yet call PAL to
calculate the angle ranges.