Updated to find angle range by stepping over elevation.
This allows the case of the nasmyth angle going back
on itself to be handled.

Also split the "parallactic angle" calculation into
a separate function, and edited it to ensure it will
return a continuous range of angles.  Checked the output
of this function visually against sourceplot plots
for several declinations via the main() method test.