Updated FTS offset angle and radii from document.
offsetAngle changed to 90 degrees so that, with Port 1 set as the
tracking port, the direction of the reference arc is from s8d to s8c.

Also made it so the reference arc display updates when FTS parameters
change by moving the resetTPE() function up from EdIterRasterObs
to OtItemEditor and invoking it in EdIterFTS2Obs.