Fixed size of circles drawn for square chop areas.
When a chop observation of a square science area is plotted
in a coordinate system other than "Tracking", the OT
uses a circular representation to indicate the annulus in which
the chopped beams should fall.  However the radius of this
circle was being incorrectly calculated (OMP fault 20111014.006)
as twice what it should have been.

Re-wrote much of the chop plotting code and introduced
a ScienceArea class in order to make the code
less repetitive.