Improving commonality between MSB and Observations.
Added a class GuiUtil and moved the createLabel convenience methods
there from IterRasterObsGUI.

Adding a MsbObsCommonGUI class as a new superclass of
ObsGUI and MsbEditorGUI and started to move code into it.

Added isMSB method to SpMSB, which always returns true.  (SpObs
already has a method which will override this.)

Allowed SpObs to use the SpMSB getTitle method, which fixes the bug
that removed observations showed a negative repeat counter.

Starting to tidy MSB and Observation GUI code.