Adding restore button for remaining count.
Between Java 1.5 and 1.6 the ActionEvent for combo boxes
appears to have changed to no longer firing when the currently
selected item is re-selected, which caused the (UN)REMOVE option
to stop working as intended.  Therefore a restore button has
been added which should appear when the observations remaining
count for an MSB or Observation is showing that it is removed
(a negative remaining count).
(cherry picked from commit 9c4fc9d16305d9f0bc68ebccaa1c8425af6c92b3)