Prioritize dialog added with set / shift options.
The dialog appears when the Prioritize button is clicked.  It offers
the previous 'automatic' behaviour as well as the options to set
or adjust the priorities by a fixed amount.

(cherry picked from commit 08038447932b2047cab1d81734c5eef3a7531328
with updates to cope with internal frames still being in the code)