Positions keep their coordinate system on drag.
When a TpePositionFeature is dragged, it should now try to convert the
coordinates as returned by the image display (FK5) to those of the object,
rather than forcing the object to use FK5.

Removed unused class TpeTargetPosFeature.

Moved automatic coordinate system conversion code from EdCompTargetList to
CoordConvert, and deleted un-used 'by hand' methods from that class.
This allowed the position feature to make use of the conversion method.