Added RxA3M as a new receiver.
As decided at the 2013/01/29 SCUBA-2 / SciCom weekly meeting.

Having this as a new receiver should allow people to load
existing programmes into the OT.

The source file EDFREQ/src/edfreq/ReceiverList.java appears to have
a list of receivers which duplicates that in acsis.cfg but
I don't understand the RxA3 entry.  Since the configuration file
overrides it, I just provided a copy of that code for RxA3M.

Note: it was necessary to have RxA3M before RxA3 in the ordering
because of the way the code pattern matches on the receiver
names in acsis.cfg.

Used the RxA3M estimated receiver.info values from the 20120810
version of HITEC.