Validate DR recipe based on the eye's instrument.
The checkDRRecipe function now takes the instrument as a
parameter so that it can check that the recipes are suitable
for that instrument.  Previously it automatically determined
the instrument by searching up the tree, but that could fail
if the instrument were defined at a higher level in the
tree structure than the DR recipe.

Also added an extra sanity check that the instrument passed
to the function is the same as the instrument found by the
previous method of searching up the tree.  If not, it's
very likely that the component has the wrong recipes in it.