EdFreq: prevent close line obscuration in popup.
The LineCatalog.returnLines() method writes details of the
spectral lines within a given frequency range to an array
indexed by the pixels of the "Emission Lines" plot.  This
is fine for generating the plot, but the same array is also
used to create the pop-up menu from which a user must
select lines.  If two lines would be plotted in the same
pixel then they overwrite eachother in this array.

In order to avoid disrupting the general manner in which
this works, this patch adds an optional parameter
"lineStoreExtra" to LineCatalog.returnLines() which
specifies a data structure into which it should store
the details of the lines which would otherwise have
overwritten other lines in the array.  The EmissionLines
class can then use this data structure to add the
additional nearby lines to the popup menu.

This is intended to fix fault: 20070830.001