Fix copyright name for Allan Brighton follows
Following an email discussion between myself and Allan Brighton
on October 5th 2013, Allan clarified that the files in the OT
bearing the header

    **
    * Title:        JSky<p>
    * Description:  <p>
    * Copyright:    Copyright (c) Allan Brighton<p>
    * Company:      <p>
    * @author Allan Brighton
    * @version 1.0
    */

had been automatically generated by JBuilder and the copyright
field was simply filled in with the name of the developer. Allan
reports that this code was written for Gemini and should have the
same copyright and license as for the other Gemini files. This
patch replaces the copyright line with AURA.

Many of these files (particularly those in the ukirt package)
are mostly not Gemini code at all but inherited the header when a new
file was begun based on a template from the Gemini tree. This patch
does not address that issue.