Add copyright statement to ATC files.
Added new copyright statement to files with ATC code
following the email below:

Date: Thu, 31 Oct 2013 14:26:10 +0000
From: "Bridger, (Alan) (ROE,UKATC)" <Alan.Bridger@stfc.ac.uk>

We have no objection to your putting the OT onto github, and no objection to
the GPL licensing. For your information we consider our licensing approach
on a case by case basis, in this case we are happy with GPL.

With regard to the ATC code in the sources: We would appreciate if you could
replace the old statement with the following:

  Copyright 2007-2013 United Kingdom Astronomy Technology Centre, an
  establishment of the Science and Technology Facilities Council.

and indeed to then insert the GPL license statement. Hope this is fine for
you.