Import catalog patches for JSky
The "Additional" directory houses modified / additional catalog
files for use with JSky.  I suspect they were originally inserted
into the JSky JAR files, but a more convenient way to include
them is to rebuild JSky from source after placing these
files in the jsky-catalog/src/main/resources directory.  It might
make more sense for the Additional directory to be a separate
branch since it's not used to actually build the source
present in the master / release branches.

This commit imports the current version of these catalog files
from the local JSky working directory, so that this repository
contains the information necessary to build JSky with the catalog
configuration currently used by the JAC OT.