Add development utilities.
These are short scripts which are useful for maintaining
the OT repository:

merge-*.sh: merge a subtree branch into the current branch.

update-*.sh: copy files into the working tree.

Note that the update scripts expect JSky / Jelly to be checked
out into the same directory as this repository is.