In TranslationUtils.recurse assign previous
The "previous" value in this method was initialized to null and only
assigned inside a (previous != null) if block.  Therefore it was
never being assigned to an actual object and so previous.translateEpilog
was never being invoked.  The assignment has been moved outside the if
block so that the calls to previous.translateEpilog are now made.
This could affect the translation of SpIterSky, this being the only
thing which appears to do anything in translateEpilog.