Add script to timestamp webstart JAR files
Some users are experiencing problems with webstart whereby they
see the error "JAR resources in JNLP file are not signed by same
certificate", depsite the fact that if the files on our server
are checked, they do appear all to be signed with the same key.
In case some kind of cache or firewall is causing this problem,
try re-naming the files to include the current date and time.