Only reset recipes when wrong for the instrument
The JCMT EdDRRecipe class previously cleared all the recipes
from the SpDRRecipe component whenever the GUI was used to look
at a component related to an instrument which was of a different
type to the instrument related to the last viewed SpDRRecipe
component.

This was the cause of the "whack-a-mole fun with DR components in
the OT" fault (20150324.011).  The corresponding science program uses
both SCUBA-2 and RxA3, and moving between the DRRecipe components in the
program caused the recipe selections to be wiped out.

This commit removes the automatic clearing of the recipes when
the instrument type related to the currently viewed DRRecipe
component changes, and replaces it with a test of whether
the recipes in the DRRecipe component are actually suitable
for the current instrument.  This test is performed regardless
of which type of instrument was previously viewed, as that
is completely irrelevant.