Actually change sideband when invalid for new freq
When a new frequency is entered which is not reachable from the
current sideband, the OT showed a dialog saying that it was changing
the sideband.  However it always just changed it back to the current
sideband.  Now it should actually change to the other sideband if the
new frequency can be reached using it.