JCMT: always write rotator angles as vector
Added a boolean method to the base "SpItem" class to identifiy
"attributes" which should be written as XML elements containing
a series of <value>...</value> elements even if there is only
one value.  Used this to ensure that the "rotatorAngles" element
is always in this form, so that it complies with the XML schema.