JCMT Validator: add check for rotator angles
Now if a restriction has been applied to the rotator angles,
issue an error if there is no scheduling constraint.  (It is
already an error to have a scheduling constraint with no elevation
information.)  If there is a scheduling constraint, instead
issue a warning to mention that this is an experimental feature.