Make OT class no longer extend JFrame
The OT class used to extend JFrame, but the JFrame which was actually
used was a JFrame constructed in OT's "main" (static) method.  (An
instance of OT was passed to OTMenuBar and used as its parent
component, but this seems incorrect as it wasn't actually used as
a component!)

The OT class constructor has been made private to prevent accidental
construction of instances of the class.