Construct OT GUI on the Swing thread
The routine to construct the OT GUI is now passed as a Runnable
to VersionSelector.checkVersions, to be called from the Swing thread
once the version configuration is complete.

VersionSelector.checkVersions now uses SwingUtilities.invokeLater
to construct its own GUI, or to call the OT GUI routine if no
configuration is required.  A second Runnable is given to be called
in the event of failure (the use closes the selection window or
no configurations are available for selection).