Make SpTreeGUI an abstract class
The methods implementing its actions are now abstract in order to
allow the compiler to detect whether any are not implemented.  Since
this class is now abstract it can't have a "main" test method which
tries to create an instance of it.  The "statusAction" has been
removed as it did nothing and doesn't appear to be used anywhere.