Don't show offsets in survey container for JCMT
The telescope position GUI contains a panel for X and Y offsets which
should only be shown for UKIRT.  It is hidden for JCMT by the constructor
of EdCompTargetList.  However when a telescope position GUI is shown
in a tabbed pane for survey containers, the tabbed pane change listener
toggles the visibility of all its components.  (I'm not sure why
this was necessary and commit b3cfd4e2ad0449c426eb6a71e8cb0a9fb54aec96
doesn't really explain.)

This commit adds extra logic to the change listener method in EdSurvey
so that it always keeps this panel hidden for JCMT.