lib: Remove Starlink signatures from jar files
Removing the old signature from the "jhbasic.jar" file prevented
Java WebStart from prompting the user to accept that particular
signature.  This reduced the number of signature prompts from 3
to 2, the remaining being for the OT package and the Starlink
jar files.  Hopefully by removing the old Starlink signature
from these files, we can eliminate the 2nd prompt so that the
user only has to accept one signature -- that used when preparing
an OT release.