Use SpInstHeterodyne's redshift calculations
EdCompInstHeterodyne now uses the SpInstHeterodyne component
to perform redshift calculations.  This is to avoid having the logic
in multiple places.  Also, since SpInstHeterodyne now recalculates
the sky frequency before saving or validating the science program,
using its logic in the editor is a good idea as it means that the
sky frequency shown on the panel is more closely related to what
will actually be written into the science program XML file.

EdCompInstHeterodyne still needs a getRedshift method (this is part
of the HeterodyneEditor interface which it implements) but this now
just calls the appropriate SpInstHeterodyne method.