Clear "previousSystem" when editing a new target
The EdCompTargetList maintains a private variable "previousSystem"
which it uses to see when a different type of conic coordinates are
used, and when that happens, to clear the values.  Unfortunately this
meant that it cleared the values when changing from one target to
another if they had different types of conic coordinates!  (See
fault 20140414.005.)  This commit adds a quick workaround, clearing
the "previousSystem" variable when a different target is edited.