Add ability to plot WFCAM AG at SKY position
See fault 20180522.002.  This adds a new TpeImageFeature class,
"TpeWfcamAutoGuiderSkyFeature" which simply extends
TpeWfcamAutoGuiderFeature and provides a different constructor.
TpeWfcamAutoGuiderFeature, if given a position type name,
now uses a similar calculation as TpeGuidePosFeature to
get the position of the SKY tag and subtracts the position of the
base tag from it to get offsets which can be applied to the
autoguider polygon to move it to the SKY location.

Caching of the polygon to plot has been removed from
TpeWfcamAutoGuiderFeature for now so that the polygon is
updated when the SKY position changes.