Remove "tag" combobox from telescope position GUI
This was disabled whenever the base tag was selected, but the box
only ever contained the base tag.  Therefore it was always disabled.
(The editor would have set the selected value to other tags, but
since only the base tag was ever added to the list of choices, that
did nothing.)

Therefore removing the box from the GUI and the associated setTag /
changeTag methods since they are never actually used.