Validator: use Format.toDouble to check PM values
Now check whether proper motion (and parallax) are specified by
using Format.toDouble and comparing to 0.0.  This is to match the
test done by SpTelescopeObsComp when deciding whether or not to
write XML elements for these values.  It now handles the case where
methods (such as SpTelescopePos.getPropMotionRA) return "0" rather
than null or the blank string.  (Previously this was seen as a
defined value rather than nothing.)