Make SoapClient error text selectable
This adds a new class "SelectableDialog" with some static methods
to display JOptionPane message dialog boxes with selectable text.
(This is done by passing JOptionPane a JTextArea rather than just
a String.)  The errors shown by the SoapClient class now make use
of this new class.

This is mostly of use for the Query Tool which uses the SoapClient
class from this package.