Allow asymmetrical het. receiver IF bandwidths
Previously we assumed that the preferred IF frequency for each
heterodyne receiver was in the center of its available bandwidth.
Thus the receiver configuration (in "acsis.cfg") only needed to
give the preferred IF frequency and one bandwidth value.

In order to accommodate new receivers where the preferred IF
frequency is not in the center of the tunable range, the IF
bandwidth in the configuration file has been separated into
lower and upper half-bandwidth values.  The tunable range is then
from (IF - lower half-bandwidth) to (IF + upper half-bandwidth).
(Note that in LSB, since a higher IF frequency corresponds to a
lower sky frequency, the "upper half-bandwidth" values applies
to the lower sky frequency end of the range.)