Don't configure freq. editor for special config.
Previously EdCompInstHeterodyne used to configure the frequency
editor when a special configuration was selected.  However it
does not appear that doing this would currently have any effect:
a special configuration can only be chosen when the frequency editor
is hidden, and opening it causes it to be reconfigured.

The "moveSlider" methods are no longer required.