Get frequency editor configuration directly
The SideBandDisplay.getCurrentConfiguration method now gets the
configuration directly via the FrequencyTable "getLineDetails"
and "getSamplers" methods, rather than by parsing the text of the
line button.  The "getLineDetails" method returns null instead
of throwing an exception in the case of the subsystem number
being out of range.  The "getLineText" method was removed.