Clamp "top sideband" only after configuring editor
This defers setting up the links from the first sideband to the
frequency editor and instrument editor, which it uses to "clamp"
the line, until after the editor has been configured.  This is
to avoid it adjusting the LO frequency during this process.

It would be nice to generally disable the frequency editor's
automatic adjustments which copying the tuning information from
the instrument component to the frequency editor.  However some
of these updates might be important if the initial tuning
information is problematic?  For now it is the LO adjustment
which seems to be causing problems with the updated way in
which the information is copied, so just this particular
step is disabled for now until configuration is complete.