Highlight selected sideband scroll bar
The SamplerWatcher "updateSamplerValues" event now includes
the sideband so that the "SideBandScrollBar" can be informed
whether or not it is the selected sidebar.  The "SideBand" class
also now sets an "isInvalid" status of the "SideBandScrollBar"
rather than directly setting its color if the bandwidth is too large.

The "SideBandScrollBar" and "SideBandUI" inner classes of
"FrequencyTable" have been changed to (static) nested classes
and the double nesting of "SideBandUI" removed.