Extract getBandSpecSelection method
Previously the logic to determine the currently available BandSpecs,
and which of those was selected, was repeated twice in the
EdCompInstHeterodyne class.  This logic has now been extracted
to create a new method in the SpInstHeterodyne class, taking
advantage of the fact that the OT is now build as a single package
so that SpInstHeterodyne can access the BandSpec information from
the edfreq.Receiver class.

The current special logic for the 3-subsystem case has been
retained in the extracted method.  Having it in just one place
will, however, allow it to be generalized later.