Use number of CMs in BandSpec selection
This adjusts the getBandSpecSelection method so that it considers
the number of CMs used for each band instead of having a fixed
special case for 3 subsystems.

In the 3-subsystem case, it is now the first subsystem which
has the greater mode selection (BandSpec "3" using 2 CMs, while
the others have BandSpec "4" using 1 CM each) rather than the
third as previously.