Clamp "top sideband" by setting LO directly
The "top sideband" is "clamped" by adjusting the LO frequency so that
it is tuned to the requested rest frequency.  This commit changes the
way this is done by computing the LO frequency directly rather than
as an incremental change from the previous value.

The code to do this has been moved from the "SideBand" class to the
"FrequencyTable" class to combine it with the logic which sets "No Line"
for the other subsystems when the clamped system's IF is changed.

After this change, the top sideband clamping should work correctly
when the IF is adjusted both by using the sliders and by entering
a frequency numerically.