Avoid chained modes starting on odd CM numbers
A comment in the JCMT Translator code mentions that the OT is supposed
to ensure that chained modes do not start on an odd CM number.  It looks
like this should not happen given the current bandwidth mode lists,
but this commit adds an extra check to avoid matching modes in this
case.