Use getBandSpecSelection to set up frequency table
The FrequencyTable class now uses the
SpInstHeterodyne.getBandSpecSelection method to get the available
bandSpecs while initally setting up the panel.

This addresses fault 20150415.004 (incorrect resolutions and bandwidth
choices in the 3-subsystem case) but only when the bandwidth choices
do not depend on the existing selections (such as the 2-subsystem case
with hybrids of more than 2 subbands).