Update freq. table bandwidth selections on change
The ItemListener for the bandwith comboboxes on the frequency table
has been changed from the Sampler class itself to an anonymous listener
in the FrequencyTable class.  This is so that the update method can
take into account the bandwith selection for all subsystems.

The Sampler class is now configured via a BandSpec rather than lists
of bandwidths and numbers of channels.  It also populates the combobox
with choices based on the BandSpec.