Construct freq. editor components with actual LO
Since the setting of the LO frequency was moved earlier in the
SideBandDisplay.updateDisplay method, the EmissionLines,
SkyTransmission and GraphScale objects were no longer receiving
the change event when the LO frequency was set.

Therefore pass the actual LO frequency and bandwidth corresponding
to the display size rather than an arbitrary value in the middle
of the tuning range to the constructors of these objects.  They
are then initially set up correctly.

Also the code to populate the data structures in the SkyTransmission
object was extracted into a new method "updatePlotData".  This is
now called both in the constructor and in the change handler method.
Previously there were copies of this code in both places, but that
in the constructor was incomplete so that the object was initially
not set up correctly.