Add system option to SpIterOffset
This adds a system control to the SpIterOffset GUI which is enabled
for JCMT only.  Since the TOML for SpIterOffset seems to use a TCS-style
obsArea element, the offset is stored as an optional SYSTEM attribute of
every OFFSET element.  When the XML is read, any such attribute sets
the iterator's system.

This feature will probably only be enabled for the local OT rather than
being included in the released version.  Therefore the default option
is blank, causing the XML to be written as before, without a SYSTEM
attribute.  Its immediate purpose is to allow a mosiac of offset
Az/El jiggles to be specified, with offsets in Az/El coordinates.