Add support for browser log in with force store
When using OAuth, since the OMP responds to a timestamp issue with a
SOAP error, we do not receive the OMP log in token despite the
one-time access code having been consumed.  Therefore if the user
wishes to force the program storage, so we need to return to
Hedwig to obtain a new access code.

The handling of SpChangedOnDiskException is now done in the SwingWorker
done method so that invokeAndWait is not needed for the "force"
dialog box.  Both cases (browser and other log in) can be handled here.