Use BandwidthOption class for selection controls
The bandwidth selection controls on the heterodyne setup panel and
in the frequency editor now use instances of a new BandwidthOption
class rather than plain string entries containing the formatted
bandwidth.  This allows additional information such as the number
of channels and the number of CMs used to be attached.  The
number of CMs used is included in the displayed text, in order
to help the user determine which CMs will be used for each
spectral region, in cases where that may matter.