Update Trx data for Namakanui
This adds new non-sideband-specific data for Uu and Aweoweo, which
is used for the noise estimates shown in some "eye" panels.  For this
data, very approximate models were created by averaging the LSB and USB
models used by the ITC (which for Uu exclude the spikes) followed by
compression for linear interpolation with 5K tolerance.

New sideband-specific data for Aweoweo has been imported from the ITC.
This is for plotting in the frequency editor.