Interpolate end points of frequency editor graphs
This improves the plotting of graphs (e.g. receiver temperature
for Aweoweo) with few points.

We now attemt to do linear interpolation to get values (of receiver
temperature and transmission) at the start and end of the graph.
This avoids gaps at the start and end if the data are sparsely
sampled.