Add class using the new Horizons API
This uses the JSON response format for requests to the new Horizons
API.  It returns the text query result, extracted from the JSON
response.  Parsing the result will be left to the existing Horizons
class which already performs this task for the previous web service.
Minor informational messages (returned alongside a response) and
API version mismatch warnings are printed to standard error.  Other
errors are raised as a nested exception class HorizonsApiException.