Read opacity data files in HeterodyneNoise.init
Now read the opacity data files once during the initialization
of the HeterodyneNoise class.  The values are now stored as
a TreeMap<Double, Double> where the filename was previously stored.
Also update the variables referring to values from this file
from "transmission" to "opacity".

This splits the previous "getTransmission" method into two
methods, "readOpacityFile" and "getOpacity".  The _availableBands
attribe is now declared private to ensure no other code is affected
by this change.