Use opacity data files for freq. editor plot
The SkyData class of the frequency editor now obtains sky transmission
data from the HeterodyneNoise class, which reads the opacity data files.
This replaces a separate table of transmission data which used to be
located in the SkyData class itself.

This now allows the frequency editor to be opened for Alaihi, since
the new opacity data files cover the relevant frequency range.