Use updated nameType for MSB title
Since the MSB title may appear in the MSBTITLE FITS header, use our
type "nameType" for it (also used for target names).  Additionally
update the pattern for this type to specify only characters valid
in FITS headers.  (Purpose of previous pattern is unclear, unless
it was to avoid particular problems with the UKIRT setHeader command.)