Add button to send coordinates via SAMP
Requested by HP.

Since the position editor already includes support for SAMP, use its
existing SAMP connection.  Unfortunately the position editor only
sets this up when it is opened, so the user will have to open the
position editor first.  There did not appear to be a method to get
non-base position coordinates with offsets applied, so a new method
was added for this, combining the logic from some existing methods.