Combine checkTargetList methods
Both SpValidation and JcmtSpValidation had checkTargetList methods,
with the latter (subclass) calling the former (superclass) method at the
end.  However the checks in the JcmtSpValidation version were almost the
same as those in the superclass (but only applied in the case
of heterodyne instruments).  Therefore combine them and remove
the subclass version.  This means applying an range of -50 ... 90
to the declination validation and adding a note about any conversion
applied.

The two versions restricted (different sections of) the checks to only
the base position, but the reason for this restriction was unclear.
Therefore remove it and allow the method to check all positions.