Partially reapply: Add system option to SpIterOffset
Re-apply this commit, adding the system option of SpIterOffset,
but leave the GUI always disabled.  This was intended to be
a temporary work-around for fault 20200305.004 but since it
is still in place, it would probably be a good idea for the
release version of the OT to be able to processes the
corresponding XML correctly.

(previously reverted in 5e7b527cdb1411a13fa5e370d0246c9f643c7cd6)

(cherry picked from commit 0ca2e7e7b9d43d0aff8af738077a7ae6906f0867)