Correct detection of blank lines in SC file
The previous test of line.trim().equals("") was applied too
late because there would already be a null pointer exception
when checking whether the tag value refers to the base position.
We can probably just assume the line is blank if there are no
tokens present when we attempt to read tag, since this should be
the first thing on the line.