Only unset transitions if molecule unset
In the EdCompInstHeterodyne._updateTransitionChoice method, if the
first molecule was unset (i.e. "No Line") then all of the transition
names would also be unset (i.e. changed to "No Line").  Now only unset
the first (corresponding to the molecule which we already know is
unset) and then check the remaining entries, unsetting each only if
its molecule is unset.  This is intended to allow for the case where
the primary subsystem is a line not listed in the catalog (and
therefore given as "No Line / No Line") but there are subsequent
subsystems with lines from the catalog.  Now it should be possible to
select them in the frequency editor and not have the transition unset
when the frequency editor is closed.